Extract waveform paint primitives
